Address review: relocate expandable tests, expand changelog, add TypeError check
- Move the six default-expandable tests from tests/unit/common/test_table.py
  to tests/unit/test_io_hdf5_h5tools.py; prune now-unused imports and restore
  PEP 8 double blank line before TestVectorIndexDtype.
- Merge TestExpandableEmpty and TestExpandableCustomList into TestExpandableArg
  with a shared fixture, and add test_bool_raises_type_error to pin the new
  TypeError contract on expandable=True/False.
- Strengthen TestDefaultExpandableWithReferences to assert the reference
  dtype and round-trip resolution of stored refs.
- Update CHANGELOG to call out the change as a breaking change and note the
  silent default layout shift for non-DynamicTable datasets.
